requirements.txt
feedparser==6.0.11
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
playwright==1.44.0
google-generativeai==0.5.4
weasyprint==62.3
Jinja2==3.1.4
APScheduler==3.10.4
python-dotenv==1.0.1
python-dateutil==2.9.0
pytz==2024.1
